move select_team_slideup and team_picker_icon files to search folder
