+++ app/screens/home/search/select_team_slideup.tsx
